requirements.txt
streamlit>=1.28.0
biopython>=1.81
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
logomaker>=0.8
plotly>=5.14.0
statsmodels>=0.14.0
ViennaRNA>=2.6.0
